Basically this commit was me doing all sorts of stuff to fix touchscreen and still failing. Fleshed out touchscreen and SPI more (doesn't work) Fixed PPU always using the engine B palette Fixed reading from KEYINPUT and EXTKEYIN Changed some debug window names Merged both memory editors into one and added a few regions Fixed DMA sending the wrong interrupts on completion Fixed palette debugger showing the wrong memory addresses Properly emulate IPC interrupt request rising edge Started making space for the firmware

+++ src/main.cpp
@@ -19,7 +19,7 @@ SDL_Scancode keymap[12] = {
 	SDL_SCANCODE_S, // Button X
 	SDL_SCANCODE_A // Button Y
 };
-u16 lastJoypad;
+u32 lastJoypad;
 
 int main(int argc, char *argv[]) {
 	if (ortin.error)
@@ -54,11 +54,12 @@ int main(int argc, char *argv[]) {
 		}
 		// Joypad inputs
 		const u8 *currentKeyStates = SDL_GetKeyboardState(nullptr);
-		u16 currentJoypad = 0;
+		u32 currentJoypad = 0;
 		for (int i = 0; i < 12; i++) {
 			if (currentKeyStates[keymap[i]])
 				currentJoypad |= 1 << i;
 		}
+		//currentJoypad |= ortin.penDown << 16;
 		if (currentJoypad != lastJoypad) {
 			ortin.nds.addThreadEvent(NDS::UPDATE_KEYS, currentJoypad);
 			lastJoypad = currentJoypad;
